requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
lightrag-hku>=0.0.12
pinecone-client>=3.0.0
numpy>=1.24.0
pydantic>=2.0.0
openai>=1.0.0
tiktoken>=0.5.0
tenacity>=8.0.0
aiohttp>=3.9.0
